refactor name & image sections as editables
